Plan and developed page has been created. Plan is called plan.html why the developed page is called home1.html
We have created the a plan.html to give the user an overview about the
mission statement, success of evaluation, targeted audience, content,
flowcharts of the website that we are going to modify/create for the
company. Furthermore, a developed page (home1.html) has been created to
give an idea to the company on how the website we are going to create
would look like.